- We switched Rekall to a new, more stable renderer model. This means you have to use new clients or Rekall will not work. The clients included in the server.deb 0.3-2 are already up to date.
- We completely removed Volatility in this release. Rekall is now
feature equivalent and just works much better so this was the
logical choice.

- We made many improvements to the SqliteDataStore, feel free to
give it a try.

- There is also a new remote data store. This datastore connects
via HTTP to one or more data servers which in turn can store the
actual data using any of the existing data stores. Currently we
support sharding between the data servers and basic rebalancing
by size of the stored data. This is still experimental but
preliminary results are very promising.

- We added labeling of clients to the gui. You can now chose some
clients from the search page, give them a label, and the run
hunts just on that group of clients. This also changes the format
how GRR stores user labels!!! This means that you have to relabel
your user accounts or use one of the console scripts we provide.

- As always there are lots of bug fixes. We fixed the race
condition which errors with "No valid status message" as an
example but there are many more.