Added CoreAccessControlManager and derivative AdminUIAccessControlManager for more granular control and logging of permissions checks
For the purpose of section 3 of the Corporate Contributor License Agreement (“Section 3”), titled “Grant of Patent License,” the “Work” is this git commit, the combination of the Contribution with the Work is git commit 5e0ab14bcc8ddf92c8775099caab8a97ab7eca21 ("Combination"), and the Contribution is therefore the difference between this git commit  and git commit 5e0ab14bcc8ddf92c8775099caab8a97ab7eca21.  Pursuant and subject to the limitations of Section 3, Yahoo grants a patent license that applies to any current or future patent claims licensable by Yahoo that would otherwise be necessarily infringed by the Combination and/or the Contribution.