Cleaned up labeling and corrected the password section to properly set blank passwords while not forcing password changes during label changes
For the purpose of section 3 of the Corporate Contributor License Agreement (“Section 3”), titled “Grant of Patent License,” the “Work” is this git commit, the combination of the Contribution with the Work is git commit 45aa4c5b21a3f4ee1fd0a4163d97aa66bd67ec2e ("Combination"), and the Contribution is therefore the difference between this git commit  and git commit 45aa4c5b21a3f4ee1fd0a4163d97aa66bd67ec2e.  Pursuant and subject to the limitations of Section 3, Yahoo grants a patent license that applies to any current or future patent claims licensable by Yahoo that would otherwise be necessarily infringed by the Combination and/or the Contribution.