Another code push. As always there are many bugfixes to many parts of the code. Notable changes:
- There are some UI improvements (more usage of AngularJS!).
- The export of results is now using softlinks to deduplicate
  data. Export is possible as .tar and .zip.
- We have more efficient collections (to store for example hunt
  results). Those are also used for hunt logs + errors so the UI
  for those should be faster for large hunts

Note that we do not ship precompiled protobuf files anymore
starting with this commit. You need to run make in the grr/proto/
directory to create those files. If you use setup.py, it will do
it for you.

Also, starting from this release, we will use Github as the main
repository and remove the one on code.google.com.