Allow server to repack template of any version.
- Fixes https://github.com/google/grr/issues/379
- Use appveyor to build windows templates on every commit.
- Convert windows template build .bat to python.
- Standardise on terminology: repacking is sticking configuration into the
template. We had used a mix of repacking and deploying to refer to this, but
usually generalised it as repacking. Making the self-extracting zip was
formerly called repacking and now has a name that describes what that method
actually does.
- Repack using the build configuration that is stored in build.yaml inside the
template. This means 3.1.5 can repack 3.1.0. The separation between building
and repacking is now stronger and clearer.
- Redefine some config variables and add new ones so we can tell what version
string comes from the current source code version.ini (Source.*) and what comes
from the build Template (Template.*). We use Source when building, Template
when repacking.
- Separate out ClientBuilder config variables where necessary: some are used
only in repacking, some only during building.
- Break repacking out into it's own library where it can be used for client
building and server initialize without having to import all the server stuff
for the client piece.
- Drop the client-side plugins system. It was more work to support this in the
repacking rewrite than it was worth.
- Fix build bugs
- Require prelink tool to at least attempt undoing any prelinking.
- Add script to install a fresh copy of grr on each build.