Prevent infinite loop when data store fails
If we get an empty or short response, the process will loop forever
trying to read the rest of the data.  Prevent this by raising an
exception if the response is not long enough.  I saw this issue on
other nodes when a data server ran out of file descriptors.