Switching Travis/Appveyor to use Python 3.
This change implies:
* Default Ubuntu version for GRR server is now Bionic, not Xenial (since
Bionic is shipped with Python 3.6).
* GRR client templates for 32 bit platforms are no longer built on
AppVeyor/Travis and are no longer included into grr-response-templates.
Building these templates after switching to Python 3 proved to be
non-trivial, and given that clients for these platforms were not
properly tested, dropping their support seems like a reasonable option.
Please note, that it's still possible to build 32-bit GRR clients, we're
just not going to support this out of the box.
* GRR version bumped to 3.4.0.0.
* A number of bugfixes made in the code.