Changed structure of general panels to rows
There is some bug (maybe a feature?) that does not allow more than 4
panels per row in a dashboard. Unfortunately there is no clear error
message so I can't be sure.
To overcome this, the general panels are now included in "general" rows
that will consist of up to 4 general panels each. And to be honest, up
to 4 panels per row make the dashboard look much cleaner.