Remove GetClientIdsFromFleetspeak
On endpoint '/search' there are two branches. One of them is where
Grafana expects the list of possible values for a variable of type
query.
In previous commits, we supported a Grafana variable that can retrieve
all possible Fleetspeak client IDs. As discussed with @mbushkov, this is
not viable for large dployments, so we decided to not support it.
The alternative to be used are Grafana variables of type textbox, which
do not query the GRRafana HTTP server for possible values, but allow for
dynamic typing of the variable value.