Added exception if requested metric is invalid
The metrics are usually fixed and are not likely to drastically change.
But, in order to not have cryptic error messages if the code does not
change but new metrics are added, then this exception is added to the
if/else branch.

Also, added a new test for this scenario.