Integrating GRR & Fleetspeak with Grafana - GRRafana HTTP server (#832)
* Simple, initial Grrafana HTTP server

* Added relevant endpoints for JSON Datasource plugin

* Implemented simple JSONRequest and JSONResponse classes

* Structured methods for all Grafana HTTP endpoints

* Fixed import error

* Defined 'e' as HTTPException

* Implemented 'on_root' method

* Return dummy metric labels to endpoint 'search'

* Moved construction of response to function fetch_available_metrics

* Returning dummy datapoints for 'query' endpoint

* Moved grrafana.py to folder /bin/

* Added context for GRRafana to work alongside GRR

With context, GRRafana HTTP server can now communicate with GRR and
Fleetspeak and use the same variables.
Also restructured some of the code in grrafana.py to work with absl
flags.

* Added GRRafana to setup.py and added entry point

* Implemented a FS util function: GetAvailableMetricsFromFleetspeak

Note that right now FS doesn't have a proto message corresponding to
GetAvailableMetricsRequest, so it will not work at the moment.

* Replaced dummy available metrics by real ones

* Implemented most of /search endpoint

* Split /search endpoint to two types

* Removed 'client_id' metric from /search endpoint results

* Fixed /search endpoint conditions

* Added GetClientIdsFromFleetspeak

* Structuring /query API endpoint

This is still not working as the Fleetspeak side is not ready, and we
still need to modify the returning data to fit our needs in format
[...,<datapoint, timestamp>,...]. Added a todo comment where applicable.

* Changed method signatures and naming

* Implemented /query API endpoint

* Fixed return on method

* Removed usages of google.protobuf.json_format.MessageToDict

* Added documentation

* Made new FS-util functions more consistent with existing functions

* Fixed err

* Changed maps() calls to list comprehensions

* Added tests for fleetspeak_utils

Added tests for all new functions added to fleetspeak_utils:
testGetAvailableMetricsFromFleetspeak, testGetClientIdsFromFleetspeak
and testFetchClientResourceUsageRecordsFromFleetspeak.

* Added relevant imports for typing etc.

* Changed the way we retrieve available metrics - we don't

After a conversation with @mbushkov, we decided to not retrieve the
available metrics from Fleetspeak but rather have them as a constant,
AVAILABLE_METRICS. The reason is that the number of metrics is (usually)
fixed, relatively small and not expected to change often, and thus
retrieving the metrics from Fleetspeak dynamically is a bit more
complicated than necessary.

* Added type annotations

* Removed leftovers from getting metrics from FS

* Removed getattr where it complicates things

* Removed unnecessary initialization of FS connector

* Changed function names for convention

* Added structure code for Grrafana object tests

* Some of changes suggested by @mbushkov and @panhania

Those changes include:
* Formatting the new added code with yapf.
* Switched all functions/methods names to PascalCase.
* Removed a list-casting that was not necessary, suggested by @panhania.
* Changed import statements (and corresponding calls) and sorted them as @mbushkov
suggested.
* Changed a return statement from Any to more specific type.
* Prefixed underscore to all methods/functions that are not used outside
grrafana module.

* Deleted function _FetchClientIds

* Changed base class of test

From test_lib.GRRBaseTest to absltest.TestCase, as suggested by
@panhania.

* Changed var name

* Fixed inaccurate usage of semi-private attribute

* Simplified client ID extraction function

* Removed unnecessary list casting

* Elaborated documentation of OnSearch method

* Made changes re @panhania's suggestions

* Catched edge case

* Remove arg 'config'

* Removed leftover references to _stub

* Fixed relevant imports for Grrafana testing

* Added initial testing for GRRafana HTTP server

This commit includes two basic tests for '/' and '/search' endpoints
(only partially for '/search') and also a structuring of the tests wth
main function.

* Added imports and added setUp

setUp method will be called before the tests methods begin, and will
initialize a client that can initiate requests to the test HTTP server.

* Added testSearchClientIds

This method tests the other branch of '/search' endpoint of the API.
This branch is initiated by Grafana when the user uses a Grafana
variable of type 'query', and is supposed to retrieve client IDs from
Fleetspeak's database.

* Added another test client

* Added test for '/query' endpoint

This test essentially mocks 2 resource usage records that return from
Fleetspeak database.
Another things in these commit - finished support for 2 test clients,
and switched to assertListEqual wherever fits.

* Removed unnecessary commented setUp code

* Testing resource_pb2.ClientResourceUsageRecord struct

The real object that is being returned from fleetspeak_utils.py, in
method FetchClientResourceUsageRecordsFromFleetspeak, is actually a list
of resource_pb2.ClientResourceUsageRecord. Therefore, the test for
'/query' endpoint now reflects that.

* Removed unnecessary imports

* Remove GetClientIdsFromFleetspeak

On endpoint '/search' there are two branches. One of them is where
Grafana expects the list of possible values for a variable of type
query.
In previous commits, we supported a Grafana variable that can retrieve
all possible Fleetspeak client IDs. As discussed with @mbushkov, this is
not viable for large dployments, so we decided to not support it.
The alternative to be used are Grafana variables of type textbox, which
do not query the GRRafana HTTP server for possible values, but allow for
dynamic typing of the variable value.

* Simplified expression in _ExtractClientIdFromVariable

* Changed from grr_afana to grr_grrafana

* Removed tests for retrieving clients' IDs

* Changed test code to use assert_called_with

* Changed constant name

* Fixed constant name

* Enforced 80-character maximum line width

* Several changes suggested by @max-vogler

Including:
* Changing constructor docstring.
* Changing naming convention of non-public API of Grrafana class.
* Changes per pystyle.
* Changing args types from List to Iterable.
* Improved accuracy of timestamp from seconds to milis.
* Import modules instead of individual members.

* Removed usage of getattr()

* Implementing a constrcutor for JSONResponse

Re @max-vogler's comment, the function json.dumps() is now used on all
JSON-serializable responses that are constructed. Moreover, the
constructor attaches the correct mimetype to all JSON responses.

* Added namedtuple 'TargetWithDatapoints' to avoid casting

* Removed unnecessary cases

* Removed Grrafana method WsgiApp

Embedded its functionality into the method __call__.

* Changed method name

* Changed endpoint variable to contain method

* Marked type definitions as private

* Added exception if requested metric is invalid

The metrics are usually fixed and are not likely to drastically change.
But, in order to not have cryptic error messages if the code does not
change but new metrics are added, then this exception is added to the
if/else branch.

Also, added a new test for this scenario.

* Probing for edge-case ClosingIterator

* Modified JSONResponse constructor

It now passes all args/kwargs after manipulation to super().

* Changed branch statement

* Added 'response' as arg to JSONResponse ctor