Created new abstract class 'Metric'
This abstract class encapsulates a single metric that can be fetched
from a FS-based GRR deployment. For instance, such a metric can be "Mean
Resident Memory MB".

This class is created so that we can introduce metrics which aren't
targeted at individual clients alone, but also multiple clients
(statistics page in GRR).